2021_09_21
1. 뒤로 가기 및 종료 시스템 구현

2. 회원 탈퇴 시 DB 동시 삭제 구현

+++ Developer_Planner_Test/app/src/main/java/com/example/login_ex/UpdateUserInfo.java
@@ -11,21 +11,27 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 
 public class UpdateUserInfo extends AppCompatActivity {
 
     AlertDialog.Builder dlg;
     DialogInterface.OnClickListener yes;
     String password = "";
+    FirebaseUser user;
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_update_user_info);
         Intent intent = getIntent();
         password = intent.getStringExtra("password");
+        user = FirebaseAuth.getInstance().getCurrentUser();
 
         findViewById(R.id.changePassword).setOnClickListener(onClickListener);
         findViewById(R.id.unregister).setOnClickListener(onClickListener);
@@ -33,19 +39,34 @@ public class UpdateUserInfo extends AppCompatActivity {
         yes = new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
-                FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
+                if (user != null) {
+                    // firebase Auth 회원 삭제
+                    user.delete()
+                            .addOnCompleteListener(new OnCompleteListener<Void>() {
+                                @Override
+                                public void onComplete(@NonNull Task<Void> task) {
+                                    if (task.isSuccessful()) {
+                                        Toast.makeText(UpdateUserInfo.this, "정상적으로 회원이 탈퇴되었습니다", Toast.LENGTH_SHORT).show();
+                                        Intent intent = new Intent(UpdateUserInfo.this, LoginActivity.class);
+                                        startActivity(intent);
 
-                user.delete()
-                        .addOnCompleteListener(new OnCompleteListener<Void>() {
-                            @Override
-                            public void onComplete(@NonNull Task<Void> task) {
-                                if (task.isSuccessful()) {
-                                    Toast.makeText(UpdateUserInfo.this,"정상적으로 회원이 탈퇴되었습니다", Toast.LENGTH_SHORT).show();
-                                    Intent intent = new Intent(UpdateUserInfo.this, LoginActivity.class);
-                                    startActivity(intent);
+                                        DatabaseReference userDB = FirebaseDatabase.getInstance().getReference();
+                                        // firebase Realtime Database 회원 정보 삭제
+                                        userDB.child("users").child(user.getUid()).removeValue()
+                                                .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                    @Override
+                                                    public void onSuccess(Void aVoid) {
+                                                    }
+                                                })
+                                                .addOnFailureListener(new OnFailureListener() {
+                                                    @Override
+                                                    public void onFailure(@NonNull Exception e) {
+                                                    }
+                                                });
+                                    }
                                 }
-                            }
-                        });
+                            });
+                }
             }
         };
